Remove broken test for empty hashed bins following delegate_hashed_bins
The test claims to "Ensure each hashed bin initially contains zero targets"
but the implementation is faulty. It is only checking for the presence of
target1_filepath, which isn't present in the delegation's target_files
list. However the same file is listed in a delegation by relative path from
the targets directory. The correct way to test that each hashed bin initially
contains zero targets would be something like:
```
for delegation in self.targets_object.delegations:
  self.assertTrue(len(delegation.target_files) == 0)
```
which is not true with the current implementation of delegate_hashed_bins()

Signed-off-by: Joshua Lock <jlock@vmware.com>