updater: reduce complexity of spec_version check
This patch significantly reduces the lines of code, and complexity of the
logic, to compare the spec_version in downloaded metadata against the
version we are expecting.

We achieve the LoC reduction by removing extraneous comments and reducing
the verbosity of comments, exception messages and log messages.

Complexity is reduced by converting the semver version strings into
tuples of ints and comparing tuple members directly.

Signed-off-by: Joshua Lock <jlock@vmware.com>